Refactored OCP rules consumer into its file
